+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `core`: Added vector catch for ARMv6-M and ARMv7-M (#1592)
   - Currently supported are HardFault and CoreReset.
 - `cli`: The run command now prints a stack trace on `HardFault` (#1592)
+- Added a simple profiler to the probe-rs cli toolkit (#1628)
+- Added MSP432E4 target (MSP432E401Y and MSP432E411Y). (#1139)
+- probe-rs-cli: added `attach` subcommand. (#1672, #1616)
 
 ### Changed